+++ PotentialPair-Emanuela.h
@@ -1,10 +1,11 @@
 // Copyright (c) 2009-2023 The Regents of the University of Michigan.
 // Part of HOOMD-blue, released under the BSD 3-Clause License.
 
+// ########## Modified by Rheoinformatic //~ [RHEOINF] ##########
+
+
 #ifndef __POTENTIAL_PAIR_H__
 #define __POTENTIAL_PAIR_H__
-
-#include "Lifetime.h"
 
 #include <iostream>
 #include <memory>
@@ -19,14 +20,6 @@
 #include "hoomd/Index1D.h"
 #include "hoomd/managed_allocator.h"
 #include "hoomd/md/EvaluatorPairLJ.h"
-
-// paniz_write
-#include "hoomd/HOOMDMPI.h"
-#include <fstream>
-#include <sstream>
-#include <vector>
-#include <math.h>
-#include <cmath> 
 
 #ifdef ENABLE_HIP
 #include <hip/hip_runtime.h>
@@ -98,7 +91,7 @@ template<class evaluator> class PotentialPair : public ForceCompute
     typedef typename evaluator::param_type param_type;
 
     //! Construct the pair potential
-    PotentialPair(std::shared_ptr<SystemDefinition> sysdef, std::shared_ptr<NeighborList> nlist, Scalar K = 0.0);
+    PotentialPair(std::shared_ptr<SystemDefinition> sysdef, std::shared_ptr<NeighborList> nlist);
     //! Destructor
     virtual ~PotentialPair();
 
@@ -121,7 +114,6 @@ template<class evaluator> class PotentialPair : public ForceCompute
     virtual void setROnPython(pybind11::tuple types, Scalar r_on);
     /// Validate that types are within Ntypes
     void validateTypes(unsigned int typ1, unsigned int typ2, std::string action);
-
 
     //! Shifting modes that can be applied to the energy
     enum energyShiftMode
@@ -192,19 +184,11 @@ template<class evaluator> class PotentialPair : public ForceCompute
         {
         return m_tail_correction_enabled;
         }
-    //S
-    void setK(Scalar K) {
-        m_K = K;
-    }
-    Scalar getK() const {
-        return m_K;
-    }
-    std::shared_ptr<Lifetime> LTIME;
-    //F
- #ifdef ENABLE_MPI
+
+#ifdef ENABLE_MPI
     //! Get ghost particle fields requested by this pair potential
     virtual CommFlags getRequestedCommFlags(uint64_t timestep);
- #endif
+#endif
 
     //! Calculates the energy between two lists of particles.
     template<class InputIterator>
@@ -221,9 +205,10 @@ template<class evaluator> class PotentialPair : public ForceCompute
     std::vector<std::string> getTypeShapeMapping() const
         {
         std::vector<std::string> type_shape_mapping(m_pdata->getNTypes());
+        unsigned int pair_typeids[2] = {0, 0}; //~ define default typeIDs as zero [RHEOINF]
         for (unsigned int i = 0; i < type_shape_mapping.size(); i++)
             {
-            evaluator eval(Scalar(0.0), Scalar(0.0), this->m_params[m_typpair_idx(i, i)]);
+            evaluator eval(Scalar(0.0), Scalar(0.0), pair_typeids, Scalar(0.0), this->m_params[m_typpair_idx(i, i)]); //~ add radcontact, array for pair_typeIDs [RHEOINF]
             type_shape_mapping[i] = eval.getShapeSpec();
             }
         return type_shape_mapping;
@@ -244,7 +229,6 @@ template<class evaluator> class PotentialPair : public ForceCompute
     protected:
     std::shared_ptr<NeighborList> m_nlist; //!< The neighborlist to use for the computation
     energyShiftMode m_shift_mode; //!< Store the mode with which to handle the energy shift at r_cut
-    Scalar m_K;
     Index2D m_typpair_idx;        //!< Helper class for indexing per type pair arrays
     GlobalArray<Scalar> m_rcutsq; //!< Cutoff radius squared per type pair
     GlobalArray<Scalar> m_ronsq;  //!< ron squared per type pair
@@ -307,9 +291,11 @@ template<class evaluator> class PotentialPair : public ForceCompute
                 {
                 for (unsigned int type_j = 0; type_j < m_pdata->getNTypes(); type_j++)
                     {
+                    unsigned int pair_typeids[2] = {0, 0}; //~ define default typeIDs as zero [RHEOINF]
                     // rho is the number density
                     Scalar rho_j = m_num_particles_by_type[type_j] / volume;
-                    evaluator eval(Scalar(0.0),
+                    evaluator eval(Scalar(0.0), Scalar(0.0), //~ add radcontact [RHEOINF] 
+                                   pair_typeids, //~ add array for pair_typeIDs [RHEOINF]
                                    h_rcutsq.data[m_typpair_idx(type_i, type_j)],
                                    m_params[m_typpair_idx(type_i, type_j)]);
                     m_external_energy += Scalar(2.0) * m_num_particles_by_type[type_i] * M_PI
@@ -331,12 +317,14 @@ template<class evaluator> class PotentialPair : public ForceCompute
                 {
                 for (unsigned int type_i = 0; type_i < m_pdata->getNTypes(); type_i++)
                     {
+                    unsigned int pair_typeids[2] = {0, 0}; //~ define default typeIDs as zero [RHEOINF]
                     // rho is the number density
                     Scalar rho_i = m_num_particles_by_type[type_i] / volume;
                     for (unsigned int type_j = 0; type_j < m_pdata->getNTypes(); type_j++)
                         {
                         Scalar rho_j = m_num_particles_by_type[type_j] / volume;
-                        evaluator eval(Scalar(0.0),
+                        evaluator eval(Scalar(0.0), Scalar(0.0), //~ add radcontact [RHEOINF] 
+                                       pair_typeids, //~ add array for pair_typeIDs [RHEOINF]
                                        h_rcutsq.data[m_typpair_idx(type_i, type_j)],
                                        m_params[m_typpair_idx(type_i, type_j)]);
                         // The pressure LRC, where
@@ -367,9 +355,8 @@ template<class evaluator> class PotentialPair : public ForceCompute
 */
 template<class evaluator>
 PotentialPair<evaluator>::PotentialPair(std::shared_ptr<SystemDefinition> sysdef,
-                                        std::shared_ptr<NeighborList> nlist,
-                                        Scalar K)
-    : ForceCompute(sysdef), m_nlist(nlist), m_shift_mode(no_shift),m_K(K),
+                                        std::shared_ptr<NeighborList> nlist)
+    : ForceCompute(sysdef), m_nlist(nlist), m_shift_mode(no_shift),
       m_typpair_idx(m_pdata->getNTypes())
     {
     m_exec_conf->msg->notice(5) << "Constructing PotentialPair<" << evaluator::getName() << ">"
@@ -377,15 +364,6 @@ PotentialPair<evaluator>::PotentialPair(std::shared_ptr<SystemDefinition> sysdef
 
     assert(m_pdata);
     assert(m_nlist);
-
-    //~ add bond_calc flag [PROCF2023]
-    if(m_K != 0.0)
-	{
-        LTIME = std::shared_ptr<Lifetime>(new Lifetime(sysdef));
-	}
-
-
-    //~
 
     GlobalArray<Scalar> rcutsq(m_typpair_idx.getNumElements(), m_exec_conf);
     m_rcutsq.swap(rcutsq);
@@ -614,11 +592,11 @@ void PotentialPair<evaluator>::setROnPython(pybind11::tuple types, Scalar r_on)
 
     \param timestep specifies the current time step of the simulation
 */
-
 template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t timestep)
     {
     // start by updating the neighborlist
     m_nlist->compute(timestep);
+
     // depending on the neighborlist settings, we can take advantage of newton's third law
     // to reduce computations at the cost of memory access complexity: set that flag now
     bool third_law = m_nlist->getStorageMode() == NeighborList::half;
@@ -637,10 +615,16 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
 
     ArrayHandle<Scalar4> h_pos(m_pdata->getPositions(), access_location::host, access_mode::read);
     ArrayHandle<Scalar> h_charge(m_pdata->getCharges(), access_location::host, access_mode::read);
+    //~ access particle diameter [RHEOINF] 
+    ArrayHandle<Scalar> h_diameter(m_pdata->getDiameters(), access_location::host, access_mode::read);
+    //~
 
     // force arrays
     ArrayHandle<Scalar4> h_force(m_force, access_location::host, access_mode::overwrite);
     ArrayHandle<Scalar> h_virial(m_virial, access_location::host, access_mode::overwrite);
+    //~ add virialxyi_ind [RHEOINF]
+    ArrayHandle<Scalar> h_virial_ind(m_virial_ind, access_location::host, access_mode::overwrite);
+    //~
 
     const BoxDim box = m_pdata->getGlobalBox();
     ArrayHandle<Scalar> h_ronsq(m_ronsq, access_location::host, access_mode::read);
@@ -649,139 +633,31 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
     PDataFlags flags = this->m_pdata->getFlags();
     bool compute_virial = flags[pdata_flag::pressure_tensor];
 
-    //S
-    ArrayHandle<Scalar> h_diameter(m_pdata->getDiameters(),
-                                   access_location::host,
-                                   access_mode::read);
-                                   
-    ArrayHandle<unsigned int> h_tag(m_pdata->getTags(),
-                                    access_location::host,
-                                    access_mode::read);
-
-    ArrayHandle<Scalar> h_current_neighbor_list( m_pdata->getParticleNList(),
-                                                    access_location::host,
-                                                    access_mode::readwrite);                        
-    assert(h_current_neighbor_list.data);
-    size_t p_neighbor_pitch = m_pdata->getParticleNList().getPitch();
-    //std::cout<<"new 2"<<std::endl;
-    // Copy the previous neighbor list to a new variable
-    int tot_particles = (int)(m_pdata->getN());  
-    std::vector<Scalar> h_previous_neighbor_temp(20 * tot_particles );
-    std::vector<Scalar> h_previous_neighbor_list;
-    if (m_K != 0.0 ){
-        //std::cout<<"new 3"<<std::endl;
-        for ( int i = 0; i < tot_particles; i++) {
-            if(h_current_neighbor_list.data[i] != -2){
-                for (size_t j = 0; j < 20; j++) { // 20 is the count of neighbors used
-                    h_previous_neighbor_temp[j * tot_particles + i] = h_current_neighbor_list.data[j * p_neighbor_pitch + i];
-                }
-            }
-        }
-        //std::cout<<"new 4"<<std::endl;
-        //std::cout<<"local: "<<tot_particles<<std::endl;
-        #ifdef ENABLE_MPI
-            if (m_sysdef->isDomainDecomposed()) {
-                // Calculate the total number of particles across all ranks
-                MPI_Allreduce(MPI_IN_PLACE,
-                            &tot_particles,
-                            1,
-                            MPI_INT,
-                            MPI_SUM,
-                            m_exec_conf->getMPICommunicator());
-            }
-        #endif
-        
-        #ifdef ENABLE_MPI
-        //std::cout<<"1"<<std::endl;
-        //std::cout<<"totals: "<<tot_particles<<std::endl;
-        unsigned int num_ranks = LTIME->num_rank;
-        // Initialize a buffer to store gathered data on the root rank   
-        std::vector<std::vector<Scalar>> gathered_previous_neighbor_lists(num_ranks);
-        std::vector<int> displacements(gathered_previous_neighbor_lists.size());
-        if (m_sysdef->isDomainDecomposed()) 
-            {   
-                // Gather data onto the root rank
-                all_gather_v(h_previous_neighbor_temp, gathered_previous_neighbor_lists, m_exec_conf->getMPICommunicator());
-                // 
-                h_previous_neighbor_list.resize(20 * tot_particles);
-                int offset = 0;
-                for (unsigned int i = 0; i < num_ranks; i++) {
-                    unsigned int ni = static_cast<int>(gathered_previous_neighbor_lists[i].size() / 20); // Number of particles in the current rank
-                    for (unsigned int p = 0; p < ni; ++p) {
-                        for (int j = 0; j < 20; j++) {
-                            h_previous_neighbor_list[j * tot_particles + offset + p] = gathered_previous_neighbor_lists[i][j * ni + p];
-                        }
-                    }
-                offset += ni;
-                }
-        }else{
-                h_previous_neighbor_list.resize(h_previous_neighbor_temp.size());
-                std::copy(h_previous_neighbor_temp.begin(), h_previous_neighbor_temp.end(), h_previous_neighbor_list.begin());
-            } 
-        #endif 
-    
-    }
-
-    // start the connected neighbors with -2
-    //memset((void*)h_current_neighbor_list.data, -2, sizeof(Scalar) * m_pdata->getParticleNList().getNumElements());
-    for (size_t i = 0; i <  m_pdata->getParticleNList().getNumElements(); ++i) {
-        h_current_neighbor_list.data[i] = -2 ;
-    }
-    // start the neighbors with tag of the particle so that the particle can be found
-    for (int i = 0; i < (int)m_pdata->getN(); i++){
-        unsigned int typei = __scalar_as_int(h_pos.data[i].w);
-        if(typei==0){
-        h_current_neighbor_list.data[i] = h_tag.data[i];
-        }
-    }
-
-    //F
     // need to start from a zero force, energy and virial
     memset((void*)h_force.data, 0, sizeof(Scalar4) * m_force.getNumElements());
     memset((void*)h_virial.data, 0, sizeof(Scalar) * m_virial.getNumElements());
-    size_t idx_pi = -1; 
+    memset((void*)h_virial_ind.data, 0, sizeof(Scalar) * m_virial_ind.getNumElements()); //~ add virialxyi_ind [RHEOINF]
+
+    //~ print shear rate [RHEOINF]
+    //Scalar shear_rate = this->m_SR;
+    //    //std::cout << shear_rate << std::endl;
+    //~
+
     // for each particle
-            
     for (int i = 0; i < (int)m_pdata->getN(); i++)
         {
-        /*
-        if (std::isnan(h_pos.data[i].z)) {
-            std::cout << "timestep:" <<timestep << std::endl;
-            std::cout<<"my rank :"<<m_exec_conf->getRank()<<std::endl;
-            std::cout << "Error: Position of particle tagi:" << h_tag.data[i] << " has NaN coordinates." << std::endl;
-            continue; // Skip processing this particle or handle the error
-        }
-        
-        std::cout<<"total particles:"<<(int)m_pdata->getN()<<std::endl;
-        std::cout<<"tagi:"<<h_tag.data[i]<<std::endl;
-        std::cout<<"previous neighbors"<<std::endl;
-        for (int j = 0; j < 20; j++) {
-            std::cout<<h_previous_neighbor_list[j * tot_particles + i ] <<",";
-        }
-        std::cout<<std::endl;
-        std::cout<<"current neighbors"<<std::endl;
-        for (int j = 0; j < 20; j++) {
-            std::cout<<h_current_neighbor_list.data[j * p_neighbor_pitch + i] <<",";
-        }
-        std::cout<<std::endl;*/
         // access the particle's position and type (MEM TRANSFER: 4 scalars)
         Scalar3 pi = make_scalar3(h_pos.data[i].x, h_pos.data[i].y, h_pos.data[i].z);
         unsigned int typei = __scalar_as_int(h_pos.data[i].w);
-        if (m_K != 0.0){
-            //find the particle in the previous time step particle neighbor list
-            if(typei == 0 ){
-                //bool wasnt_found = true;
-                for(int p_i = 0; p_i < tot_particles; p_i++){
-                    if (h_previous_neighbor_list[p_i] == h_tag.data[i]) {
-                        idx_pi = p_i;
-                        //wasnt_found = false;
-                        break;
-                    }else{idx_pi = -3;}
-                }
-            }
-        }
+
         // sanity check
         assert(typei < m_pdata->getNTypes());
+
+        //~ access diameter (if needed) removed in v4 upgrade, readded by [RHEOINF]
+        Scalar di = Scalar(0.0);
+        if (evaluator::needsDiameter())
+           di = h_diameter.data[i];
+        //~
 
         // access charge (if needed)
         Scalar qi = Scalar(0.0);
@@ -797,6 +673,7 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
         Scalar virialyyi = 0.0;
         Scalar virialyzi = 0.0;
         Scalar virialzzi = 0.0;
+        Scalar virialxyi_ind = 0.0; //~ add virialxyi_ind [RHEOINF]
 
         // loop over all of the neighbors of this particle
         const size_t myHead = h_head_list.data[i];
@@ -815,6 +692,16 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
             unsigned int typej = __scalar_as_int(h_pos.data[j].w);
             assert(typej < m_pdata->getNTypes());
 
+            //~ store the typeIDs of the current pair [RHEOINF]
+            unsigned int pair_typeids[2] = {typei, typej};
+            //~
+
+            //~ access diameter (if needed) removed in v4 upgrade, readded by [RHEOINF]
+            Scalar dj = Scalar(0.0);
+            if (evaluator::needsDiameter())
+                dj = h_diameter.data[j];
+            //~
+
             // access charge (if needed)
             Scalar qj = Scalar(0.0);
             if (evaluator::needsCharge())
@@ -825,6 +712,10 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
 
             // calculate r_ij squared (FLOPS: 5)
             Scalar rsq = dot(dx, dx);
+
+            //~ calculate the center-center distance equal to particle-particle contact (AKA r0) [RHEOINF]
+            Scalar radcontact = Scalar(0.5) * (h_diameter.data[i] + h_diameter.data[j]);
+            //~
 
             // get parameters for this type pair
             unsigned int typpair_idx = m_typpair_idx(typei, typej);
@@ -849,90 +740,16 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
             // compute the force and potential energy
             Scalar force_divr = Scalar(0.0);
             Scalar pair_eng = Scalar(0.0);
-            evaluator eval(rsq, rcutsq, param);
+
+            evaluator eval(rsq, radcontact, pair_typeids, rcutsq, param); //~ add radcontact, pair_typeIDs [RHEOINF]
+            //~ add diameter (if needed) removed in v4, readded [RHEOINF]
+            if (evaluator::needsDiameter())
+                eval.setDiameter(di, dj);
+            //~
             if (evaluator::needsCharge())
                 eval.setCharge(qi, qj);
 
             bool evaluated = eval.evalForceAndEnergy(force_divr, pair_eng, energy_shift);
-            //S
-            unsigned int tagi = h_tag.data[i];
-            unsigned int tagj = h_tag.data[j];
-            
-            if (m_K != 0.0){
-                if (typei == typej)
-                {   
-                    Scalar rsq_root = std::sqrt(rsq) - Scalar(2.0);
-                    if (rsq_root < Scalar(2.0))
-                    {
-                        // Check and update neighbors for particle i
-                        size_t idx_i = 0;
-                        bool not_saved = true;
-                        while (idx_i < 20){
-                        
-                            // Check if tag[j] already exists
-                            if (h_current_neighbor_list.data[idx_i * p_neighbor_pitch + i] == h_tag.data[j])
-                            {
-                                not_saved = false;
-                                break; 
-                            }
-
-                            // If current entry is empty, save tag[j]
-                            if (h_current_neighbor_list.data[idx_i * p_neighbor_pitch + i] == -2)
-                            {
-                                h_current_neighbor_list.data[idx_i * p_neighbor_pitch + i] = h_tag.data[j];
-                                break; 
-                            }
-
-                            idx_i++; 
-                        }
-
-                        if (not_saved)
-                        {
-                            // Check and update neighbors for particle j
-                            size_t idx_j = 0;
-                            while (idx_j < 20)
-                            {
-                                // Check if tag[i] already exists
-                                if (h_current_neighbor_list.data[idx_j * p_neighbor_pitch + j] == h_tag.data[i])
-                                {
-                                    break; 
-                                }
-
-                                // If current entry is empty, save tag[i]
-                                if (h_current_neighbor_list.data[idx_j * p_neighbor_pitch + j] == -2)
-                                {
-                                    h_current_neighbor_list.data[idx_j * p_neighbor_pitch + j] = h_tag.data[i];
-                                    break; 
-                                }
-
-                                idx_j++; 
-                            }
-                        }
-
-                            
-                            // save angles 
-                            bool new_connection = true;
-                            size_t is_idx = 0;
-                            while ( is_idx < 20)
-                            {
-                                if (h_previous_neighbor_list[is_idx * tot_particles + idx_pi] == h_tag.data[j])
-                                {
-                                    new_connection = false;
-                                    break;
-                                }
-                                is_idx++;
-                            }
-                            
-                        
-                    }
-                }
-            }
-             //F
-
-
-
-
-
 
             if (evaluated)
                 {
@@ -978,7 +795,9 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
                     virialyyi += force_div2r * dx.y * dx.y;
                     virialyzi += force_div2r * dx.y * dx.z;
                     virialzzi += force_div2r * dx.z * dx.z;
+                    virialxyi_ind += force_div2r * dx.x * dx.y; //~ add virialxyi_ind [RHEOINF]
                     }
+
                 // add the force to particle j if we are using the third law (MEM TRANSFER: 10
                 // scalars / FLOPS: 8) only add force to local particles
                 if (third_law && j < m_pdata->getN())
@@ -988,12 +807,6 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
                     h_force.data[mem_idx].y -= dx.y * force_divr;
                     h_force.data[mem_idx].z -= dx.z * force_divr;
                     h_force.data[mem_idx].w += pair_eng * Scalar(0.5);
-
-                    if ((timestep-100000) % 10000 == 0){
-                    Scalar fj_mag = std::sqrt(h_force.data[mem_idx].x * h_force.data[mem_idx].x 
-                    + h_force.data[mem_idx].y * h_force.data[mem_idx].y + h_force.data[mem_idx].z * h_force.data[mem_idx].z);
-                    std::cout << h_force.data[mem_idx].w << "," << fj_mag << "," << (rsq-Scalar(2.0)) << std::endl;
-                    }
                     if (compute_virial)
                         {
                         h_virial.data[0 * m_virial_pitch + mem_idx] += force_div2r * dx.x * dx.x;
@@ -1002,6 +815,7 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
                         h_virial.data[3 * m_virial_pitch + mem_idx] += force_div2r * dx.y * dx.y;
                         h_virial.data[4 * m_virial_pitch + mem_idx] += force_div2r * dx.y * dx.z;
                         h_virial.data[5 * m_virial_pitch + mem_idx] += force_div2r * dx.z * dx.z;
+                        h_virial_ind.data[0 * m_virial_ind_pitch + mem_idx] += force_div2r * dx.x * dx.y; //~ add virialxyi_ind [RHEOINF]
                         }
                     }
                 }
@@ -1021,198 +835,9 @@ template<class evaluator> void PotentialPair<evaluator>::computeForces(uint64_t 
             h_virial.data[3 * m_virial_pitch + mem_idx] += virialyyi;
             h_virial.data[4 * m_virial_pitch + mem_idx] += virialyzi;
             h_virial.data[5 * m_virial_pitch + mem_idx] += virialzzi;
+            h_virial_ind.data[0 * m_virial_ind_pitch + mem_idx] += virialxyi_ind; //~ add virialxyi_ind [RHEOINF]
             }
         }
-
-if (m_K != 0.0){
-    #ifdef ENABLE_MPI
-
-
-    //~ add bond_calc [PROCF2023] 
-    
-    	LTIME->updatebondtime(timestep);
-
-    //
-    #endif
-    for (int i = 0; i < (int)m_pdata->getN(); i++){
-        unsigned int typei = __scalar_as_int(h_pos.data[i].w);
-        Scalar3 pi = make_scalar3(h_pos.data[i].x, h_pos.data[i].y, h_pos.data[i].z);
-        if (typei==0) {
-                //Scalar kk =0;
-                size_t nonzero_count = 0;
-                unsigned int tagi = h_tag.data[i]; 
-                for (size_t idx_i = 1; idx_i < 20 ; ++idx_i)
-                {
-                    // check if i has more than 1 neighbor
-                    if (h_current_neighbor_list.data[idx_i* p_neighbor_pitch +i] != -2 
-                    && h_current_neighbor_list.data[idx_i* p_neighbor_pitch +i] != tagi)
-                    {
-                        nonzero_count++; //count the neighbors
-                    }
-                }
-                if (nonzero_count > 1)
-                {
-                    #define SMALL Scalar(0.001)
-                    // go over each angle between two different neighbors of i and apply angle force 
-                    // find the first neighbor a
-                    for (size_t idx_a = 1; idx_a < (nonzero_count); ++idx_a)
-                    {
-                        unsigned int taga = static_cast<unsigned int>(h_current_neighbor_list.data[idx_a* p_neighbor_pitch +i]);
-                        Scalar3 pa;
-                        int a = -1;
-                        //bool f5=false;
-                        for (int M = 0; M < (int)(m_pdata->getN()+ m_pdata->getNGhosts()); M++)
-                        {
-                            if (h_tag.data[M] == taga) {
-                            // Access the position of the particle with the matching tag
-                                pa = make_scalar3(h_pos.data[M].x, h_pos.data[M].y, h_pos.data[M].z);
-                                a = M; 
-                                //f5=true;
-                                break; 
-                                }
-                        }
-
-                        unsigned int typea = __scalar_as_int(h_pos.data[a].w);
-                        // find the second neighbor b
-                        for (size_t idx_b = (idx_a+1) ; idx_b < (nonzero_count+1); ++idx_b)
-                        {
-                            unsigned int tagb = static_cast<unsigned int>(h_current_neighbor_list.data[idx_b* p_neighbor_pitch +i]);
-                            Scalar3 pb;
-                            int b = -1;
-                            //bool f6=false;
-                            for (int N = 0; N < (int)(m_pdata->getN()+ m_pdata->getNGhosts()); N++)
-                            {
-                                if (h_tag.data[N] == tagb) {
-                                    // Access the position of the particle with the matching tag
-                                    pb = make_scalar3(h_pos.data[N].x, h_pos.data[N].y, h_pos.data[N].z);
-                                    b = N;
-                                    //f6 = true;
-                                    break; 
-                                    }
-
-                            }
-                            unsigned int typeb = __scalar_as_int(h_pos.data[b].w);
-
-                            
-
-                            if (tagb != taga  && taga != tagi && tagb != tagi && typea==0 && typeb==0)
-                            {
-                                // Calculate angle index
-                                unsigned int vari = tagi ;
-                                unsigned int vara = taga ;
-                                unsigned int varb = tagb ;
-
-                                unsigned int var1 = vari;
-                                unsigned int var3 = std::max({vara, varb});
-                                unsigned int var2 = std::min({vara, varb});
-
-                                unsigned int n = LTIME->num_solvent; 
-                                unsigned int current_angle_index = (var1 * n*(n-1)/2) + (2*var2*n - var2*var2 + 2*var3 - 3*var2 -2)/2;
-
-                                Scalar3 ria = make_scalar3(pa.x - pi.x, pa.y - pi.y, pa.z - pi.z);
-                                Scalar3 rib = make_scalar3(pb.x - pi.x, pb.y - pi.y, pb.z - pi.z);
-                                ria = box.minImage(ria);
-                                rib = box.minImage(rib);
-
-                                // Calculate magnitudes of vectors
-                                Scalar ria_mag = std::sqrt(ria.x * ria.x + ria.y * ria.y + ria.z * ria.z);
-                                Scalar rib_mag = std::sqrt(rib.x * rib.x + rib.y * rib.y + rib.z * rib.z);
-
-                                // Calculate dot product
-                                Scalar dot_product = ria.x * rib.x + ria.y * rib.y + ria.z * rib.z;
-
-                                // Calculate cosine of the angle 
-                                Scalar current_cos_theta = dot_product / (ria_mag * rib_mag);
-                                
-
-                                if (current_cos_theta > 1.0)
-                                    current_cos_theta = 1.0;
-                                if (current_cos_theta < -1.0)
-                                    current_cos_theta = -1.0;
-                                
-                                Scalar current_sin_theta = std::sqrt(1.0 - current_cos_theta * current_cos_theta);
-                                if (current_sin_theta < SMALL)
-                                    current_sin_theta = SMALL;
-                            
-                                current_sin_theta = 1.0 / current_sin_theta;
-                                
-                                // Calculate force magnitude for Emanuela equation
-                                Scalar B = 67.27;
-                                Scalar w = 0.3;
-                                Scalar eq_theta = 65.0 * M_PI / 180.0;
-
-                                Scalar Aia = pow(ria_mag/2, -10) * pow(1 - pow(ria_mag / 4, 10), 2);
-                                Scalar Aib = pow(rib_mag/2, -10) * pow(1 - pow(rib_mag / 4, 10), 2);
-
-                                Scalar exp_theta = exp(-pow((current_cos_theta - cos(eq_theta)) / w, 2));
-
-                                Scalar vab = (-2 * B * current_sin_theta / pow(w, 2)) * Aia * Aib * exp_theta * (current_cos_theta - cos(eq_theta));
-                                
-
-                                Scalar a11 = vab * current_cos_theta / (ria_mag * ria_mag);
-                                Scalar a12 = -vab / (ria_mag * rib_mag);
-                                Scalar a22 = vab * current_cos_theta / (rib_mag * rib_mag);
-
-                                // Calculate forces
-                                Scalar3 fia = make_scalar3(a11 * ria.x + a12 * rib.x,
-                                                            a11 * ria.y + a12 * rib.y,
-                                                            a11 * ria.z + a12 * rib.z);
-
-                                Scalar3 fib = make_scalar3(a22 * ria.x + a12 * rib.x,
-                                                            a22 * ria.y + a12 * rib.y,
-                                                            a22 * ria.z + a12 * rib.z);
-
-
-                                // compute the energy, for each atom in the angle for Emanuela equation
-                                Scalar angle_eng = (B * Aia * Aib * exp_theta)/3;
-
-                                Scalar angle_virial[6];
-                                angle_virial[0] = Scalar(1. / 3.) * (ria.x * fia.x + rib.x * fib.x);
-                                angle_virial[1] = Scalar(1. / 3.) * (ria.y * fia.x + rib.y * fib.x);
-                                angle_virial[2] = Scalar(1. / 3.) * (ria.z * fia.x + rib.z * fib.x);
-                                angle_virial[3] = Scalar(1. / 3.) * (ria.y * fia.y + rib.y * fib.y);
-                                angle_virial[4] = Scalar(1. / 3.) * (ria.z * fia.y + rib.z * fib.y);
-                                angle_virial[5] = Scalar(1. / 3.) * (ria.z * fia.z + rib.z * fib.z);
-
-                                
-                                // Update forces and virials for particle i, a and b
-                                
-                                if (a < (int)(m_pdata->getN()+ m_pdata->getNGhosts())) {
-                                    h_force.data[a].x += fia.x;
-                                    h_force.data[a].y += fia.y;
-                                    h_force.data[a].z += fia.z;
-                                    h_force.data[a].w += angle_eng;
-                                    for (int l = 0; l < 6; l++)
-                                        h_virial.data[l * m_virial_pitch + a] += angle_virial[l]; 
-                                }
-                                
-                                if (i < (int)m_pdata->getN()) {
-                                    h_force.data[i].x -= fia.x + fib.x; 
-                                    h_force.data[i].y -= fia.y + fib.y;
-                                    h_force.data[i].z -= fia.z + fib.z;
-                                    h_force.data[i].w += angle_eng;
-                                    for (int l = 0; l < 6; l++)
-                                        h_virial.data[l * m_virial_pitch + i] += angle_virial[l]; 
-                                }
-
-                                if (b < (int)(m_pdata->getN()+ m_pdata->getNGhosts())) {
-                                    h_force.data[b].x += fib.x;
-                                    h_force.data[b].y += fib.y;
-                                    h_force.data[b].z += fib.z;
-                                    h_force.data[b].w += angle_eng;
-                                    for (int l = 0; l < 6; l++)
-                                        h_virial.data[l * m_virial_pitch + b] += angle_virial[l]; 
-                                }
-              
-                            }
-                        }
-                    }
-                } 
-            }
-        }
-  
-    //F   
-    }
 
     computeTailCorrection();
     }
@@ -1227,6 +852,13 @@ CommFlags PotentialPair<evaluator>::getRequestedCommFlags(uint64_t timestep)
 
     if (evaluator::needsCharge())
         flags[comm_flag::charge] = 1;
+
+    flags[comm_flag::diameter] = 1; //~ make sure diameter is accessible in MPI
+
+    //~ add diameter (if needed) removed in v4 re-added [RHEOINF]
+    if (evaluator::needsDiameter())
+        flags[comm_flag::diameter] = 1;
+    //~
 
     flags |= ForceCompute::getRequestedCommFlags(timestep);
 
@@ -1278,15 +910,6 @@ inline void PotentialPair<evaluator>::computeEnergyBetweenSets(InputIterator fir
 
         // reset the old flags
         m_comm->setFlags(old_flags);
-
-
-        // Paniz
-        CommFlags flags = CommFlags(0);
-        flags[comm_flag::position] = 1;
-        flags[comm_flag::net_force] = 1;
-        flags[comm_flag::net_virial] = 1;
-        flags[comm_flag::particle_n_list] = 1;
-        flags[comm_flag::tag] = 1;
         }
 #endif
 
@@ -1297,6 +920,9 @@ inline void PotentialPair<evaluator>::computeEnergyBetweenSets(InputIterator fir
                                       access_location::host,
                                       access_mode::read);
     ArrayHandle<Scalar> h_charge(m_pdata->getCharges(), access_location::host, access_mode::read);
+    //~ access particle diameter [RHEOINF] 
+    ArrayHandle<Scalar> h_diameter(m_pdata->getDiameters(), access_location::host, access_mode::read);
+    //~
 
     const BoxDim box = m_pdata->getGlobalBox();
     ArrayHandle<Scalar> h_ronsq(m_ronsq, access_location::host, access_mode::read);
@@ -1315,6 +941,12 @@ inline void PotentialPair<evaluator>::computeEnergyBetweenSets(InputIterator fir
 
         // sanity check
         assert(typei < m_pdata->getNTypes());
+
+        //~ access diameter (if needed) removed in v4, re-added [RHEOINF]
+        Scalar di = Scalar(0.0);
+        if (evaluator::needsDiameter())
+            di = h_diameter.data[i];
+        //~
 
         // access charge (if needed)
         Scalar qi = Scalar(0.0);
@@ -1336,6 +968,16 @@ inline void PotentialPair<evaluator>::computeEnergyBetweenSets(InputIterator fir
             unsigned int typej = __scalar_as_int(h_pos.data[j].w);
             assert(typej < m_pdata->getNTypes());
 
+            //~ store the typeIDs of the current pair [RHEOINF]
+            unsigned int pair_typeids[2] = {typei, typej};
+            //~
+
+            //~ access diameter (if needed) removed in v4 upgrade, readded by [RHEOINF]
+            Scalar dj = Scalar(0.0);
+            if (evaluator::needsDiameter())
+                dj = h_diameter.data[j];
+            //~
+
             // access charge (if needed)
             Scalar qj = Scalar(0.0);
             if (evaluator::needsCharge())
@@ -1346,6 +988,10 @@ inline void PotentialPair<evaluator>::computeEnergyBetweenSets(InputIterator fir
 
             // calculate r_ij squared (FLOPS: 5)
             Scalar rsq = dot(dx, dx);
+
+            //~ calculate the center-center distance equal to particle-particle contact (AKA r0) [RHEOINF]
+            Scalar radcontact = Scalar(0.5) * (h_diameter.data[i] + h_diameter.data[j]);
+            //~
 
             // get parameters for this type pair
             unsigned int typpair_idx = m_typpair_idx(typei, typej);
@@ -1370,7 +1016,12 @@ inline void PotentialPair<evaluator>::computeEnergyBetweenSets(InputIterator fir
             // compute the force and potential energy
             Scalar force_divr = Scalar(0.0);
             Scalar pair_eng = Scalar(0.0);
-            evaluator eval(rsq, rcutsq, param);
+
+            evaluator eval(rsq, radcontact, pair_typeids, rcutsq, param); //~ add radcontact, pair_typeIDs [RHEOINF]
+            //~ add diameter (if needed), removed in v4 but re-added [RHEOINF]
+            if (evaluator::needsDiameter())
+                eval.setDiameter(di, dj);
+            //~
             if (evaluator::needsCharge())
                 eval.setCharge(qi, qj);
 
@@ -1452,7 +1103,7 @@ template<class T> void export_PotentialPair(pybind11::module& m, const std::stri
     pybind11::class_<PotentialPair<T>, ForceCompute, std::shared_ptr<PotentialPair<T>>>
         potentialpair(m, name.c_str());
     potentialpair
-        .def(pybind11::init<std::shared_ptr<SystemDefinition>, std::shared_ptr<NeighborList>, Scalar>())
+        .def(pybind11::init<std::shared_ptr<SystemDefinition>, std::shared_ptr<NeighborList>>())
         .def("setParams", &PotentialPair<T>::setParamsPython)
         .def("getParams", &PotentialPair<T>::getParams)
         .def("setRCut", &PotentialPair<T>::setRCutPython)
@@ -1465,8 +1116,7 @@ template<class T> void export_PotentialPair(pybind11::module& m, const std::stri
         .def_property("tail_correction",
                       &PotentialPair<T>::getTailCorrectionEnabled,
                       &PotentialPair<T>::setTailCorrectionEnabled)
-        .def("computeEnergyBetweenSets", &PotentialPair<T>::computeEnergyBetweenSetsPythonList)
-        .def_property("K", &PotentialPair<T>::getK, &PotentialPair<T>::setK);
+        .def("computeEnergyBetweenSets", &PotentialPair<T>::computeEnergyBetweenSetsPythonList);
     }
 
     } // end namespace detail
